Reimplement orbit shape in terms of e,p instead of e,a.
The problem with e,a is a singularity at parabolas. As e approaches 1.0,
a approaches infinity. And at e=1.0, a _must_ be infinity. This means
that it's impossible to specify the size of the parabola.

A better alternative is the semi-latus rectum, p, which specifes the
altitude at 90 degrees from periapsis, which is always finite and is
measured perpendicular to the effects of eccentricity.