Implement a GPU-accelerated 3D ellipse rasterizer.
After a long weekend of calculus trying to
find a closed-form solution to raycasting
an ellipse, I ended up with an irreducible
(as far as I know) quartic. So I gave up
with that approach and I've instead implemented
Newton's method to find the minimal
distance, along with some simple (but
not yet perfect) early culling strategies.